Support multiple outstanding authentication requests
Simply queue up authentication requests and process them one by
one. We could have done this in parallel but that's probably not a
very good user experience.

This feature was requested here

 https://bugzilla.redhat.com/show_bug.cgi?id=526053

and while it's a sign of bad design to be in a situation where
multiple outstanding requests are pending, this is not the right place
to fix it.

One example of this feature is that you can now run 'pkexec bash' from
two different terminals and you get two authentication requests right
after each other.